Tests: re-organize directory
Put all test classes in their new location as per issue 648.
